soap will now use all of the arguments to make out a path.
I have also set the value of the shell arguments to a local environment
variable instead in order to get rid of the escape quote function that was
previously there.

Modified the Makefile so that it generates a backup of the previous xdg-open
script which soap relies on if no regex pair matches.

string.h has been added again